+++ lib/rails/forward_compatible_controller_tests.rb
@@ -58,15 +58,10 @@ module Rails
           xhr = request_params.delete(:xhr)
           request_format = request_params.delete(:format)
           if request_params[:params].is_a?(Hash)
-            request_params.merge!(request_params.delete(:params) || {})
             request_session = request_params.delete(:session) || request_session if controller_test
             request_headers = request_params.delete(:headers) || request_headers unless controller_test
             request_flash = request_params.delete(:flash) || request_flash if controller_test
-          elsif request_params.key?(:params)
-            request_flash = request_params[:flash] if controller_test
-            request_session = request_params[:session] if controller_test
-            request_headers = request_params[:headers] unless controller_test
-            request_params = request_params[:params]
+            request_params.merge!(request_params.delete(:params) || {})
           elsif request_params.key?(:headers) && !controller_test
             request_flash = nil
             request_session = nil